+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/multiblocks/AbstractSmeltery.java
@@ -30,8 +30,8 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
  */
 abstract class AbstractSmeltery extends MultiBlockMachine {
 
-    public AbstractSmeltery(Category category, SlimefunItemStack item, ItemStack[] recipe, ItemStack[] machineRecipes, BlockFace trigger) {
-        super(category, item, recipe, machineRecipes, trigger);
+    public AbstractSmeltery(Category category, SlimefunItemStack item, ItemStack[] recipe, BlockFace trigger) {
+        super(category, item, recipe, trigger);
     }
 
     @Override